pyproject.toml
[project]
name = "telegram-lols-kicker"
version = "0.1.0"
description = "A Telegram bot that can kick people that are detected as spammers in the LOLS database"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.16.0",
    "fastapi>=0.115.6",
    "modal>=0.73.172",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.8.4",
]
